Refactoring EPUB builders: split common codes from EPUB2 builder
